+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-05 23:11:05
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-07-06 11:25:45
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '2620';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-07-05T23:11:05+05:00');
+  final String commitRef = '2625';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-07-06T11:25:45+05:00');
 }